pyproject.toml
[project]
name = "boilerplate-fast-api"
version = "0.1.0"
description = "Boilerplate FastAPI application with SQLAlchemy"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Jakub Przybylo", email = "jkb.przybylo@gmail.com" }]
requires-python = ">=3.12,<3.14"
keywords = ["fastapi", "sqlalchemy", "api", "boilerplate"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.13",
    "Framework :: FastAPI",
]
dependencies = [
    "fastapi>=0.120.1",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.4",
    "sqlalchemy[asyncio,mypy]>=2.0.36",
    "alembic>=1.14.0",
    "psycopg[binary,pool]>=3.2.0",
    "result>=0.17.0",
    "pydantic-settings>=2.11.0",
    "asgi-correlation-id>=4.3.4",
    "pydantic-ai-slim[openai]>=1.85.0",
]

[project.urls]
Homepage = "https://github.com/jprzybylo/boilerplate-fast-api"
Repository = "https://github.com/jprzybylo/boilerplate-fast-api"
Issues = "https://github.com/jprzybylo/boilerplate-fast-api/issues"

[dependency-groups]
dev = ["ruff>=0.8.0", "mypy>=1.13.0", "pre-commit>=4.0.1", "ty>=0.0.12"]
test = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.1",
    "aiosqlite>=0.22.1",
]
types = ["types-requests>=2.32.0.20241016"]

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple/"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # Pyflakes (unused imports, undefined names)
    "W",      # pycodestyle warnings
    "B",      # flake8-bugbear (likely bugs and design problems)
    "C90",    # mccabe complexity
    "I",      # isort (import sorting)
    "N",      # pep8-naming (naming conventions)
    "UP",     # pyupgrade (modern Python syntax)
    "ANN",    # flake8-annotations (type annotations)
    "ASYNC",  # flake8-async (async/await best practices)
    "S",      # flake8-bandit (security issues)
    "BLE",    # flake8-blind-except (blind exception catching)
    "A",      # flake8-builtins (builtin shadowing)
    "COM",    # flake8-commas (trailing commas)
    "DTZ",    # flake8-datetimez (datetime timezone awareness)
    "T10",    # flake8-debugger (debugger statements)
    "T20",    # flake8-print (print statements)
    "PT",     # flake8-pytest-style (pytest best practices)
]
ignore = ["ANN101", "ANN102", "S101", "COM812", "B008"]

[tool.ruff.lint.mccabe]
max-complexity = 18

[tool.ruff.lint.isort]
known-first-party = ["src", "integration"]
force-sort-within-sections = true


[tool.mypy]
python_version = "3.12"
plugins = ["sqlalchemy.ext.mypy.plugin", "pydantic.mypy"]
strict = true
show_error_codes = true
exclude = "migrations/"
explicit_package_bases = true

[tool.ty.src]
exclude = ["**/*.pyi"]

[tool.ty.rules]
# Equivalent to mypy's warn_redundant_casts
redundant-cast = "warn"
# Equivalent to mypy's warn_unused_ignores (ty has this as error by default)
unused-ignore-comment = "ignore"
# Additional helpful rules for stricter checking
invalid-ignore-comment = "error"
ignore-comment-unknown-rule = "error"
deprecated = "warn"

# Enable all core type checking rules at error level
invalid-assignment = "error"
invalid-argument-type = "ignore"  # Protocol not working
invalid-return-type = "error"
unresolved-reference = "error"
unresolved-import = "error"
unresolved-attribute = "error"
call-non-callable = "error"
missing-argument = "error"
invalid-method-override = "ignore"
invalid-parameter-default = "error"
